+++ app/src/main/java/com/theunquenchedservant/granthornersbiblereadingsystem/utilities/DoneReceiver.kt
@@ -5,7 +5,7 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.widget.Toast
-import com.theunquenchedservant.granthornersbiblereadingsystem.Marker.markAll
+import com.theunquenchedservant.granthornersbiblereadingsystem.utilities.Marker.markAll
 import com.theunquenchedservant.granthornersbiblereadingsystem.utilities.SharedPref.getIntPref
 
 class DoneReceiver : BroadcastReceiver() {
